refactor: introduce `WalletRepository` interface and abstract data source
This commit refactors the wallet repository logic by introducing a domain-level interface and migrating the implementation to use a data source pattern. This improves abstraction and ensures the repository returns domain models instead of DTOs.

- **WalletRepository.kt (Domain)**: Introduced a new interface defining the contract for wallet operations, including `getWallets`, `createWallet`, and preference retrieval.
- **WalletRepositoryImpl.kt (Data)**:
    - Renamed from `WalletRepository` and implemented the new `WalletRepository` interface.
    - Replaced direct `KtorClient` dependency with `WalletDataSource` to abstract remote API calls.
    - Updated `getWallets` and `createWallet` to map DTO response data into domain-level `Wallet` objects.
    - Updated method signatures to return `Result` types containing domain models instead of raw API responses.
- **DTO and Imports**: Updated package references for `PostWalletRequest` and related data structures to align with the new data source architecture.
