--- shared/src/commonMain/kotlin/id/my/sendiko/fintrack/wallet/core/data/WalletRepository.kt
@@ -1,41 +1,0 @@
-package id.my.sendiko.fintrack.wallet.core.data
-
-import id.my.sendiko.fintrack.core.network.KtorClient
-import id.my.sendiko.fintrack.core.network.utils.DataError
-import id.my.sendiko.fintrack.core.network.utils.Result
-import id.my.sendiko.fintrack.core.preferences.PreferencesRepositoryImpl
-import id.my.sendiko.fintrack.wallet.core.data.dto.getdetails.GetWalletsResponse
-import id.my.sendiko.fintrack.wallet.create.data.PostWalletRequest
-import id.my.sendiko.fintrack.wallet.create.data.PostWalletResponse
-import id.my.sendiko.fintrack.wallet.core.domain.Wallet
-import kotlinx.coroutines.flow.Flow
-
-class WalletRepository(
-    val ktorClient: KtorClient,
-    val prefs: PreferencesRepositoryImpl
-) {
-
-    suspend fun getWallets(): Result<GetWalletsResponse, DataError.Remote> {
-        return ktorClient.getWallets()
-    }
-
-    suspend fun createWallet(userId: String, wallet: Wallet): Result<PostWalletResponse, DataError.Remote> {
-        val request = PostWalletRequest(
-            balance = wallet.amount,
-            purpose = wallet.purpose,
-            name = wallet.name,
-            type = wallet.type,
-            userId = userId,
-        )
-        return ktorClient.postWallet(request)
-    }
-
-    fun getUserId(): Flow<String> {
-        return prefs.getUserId()
-    }
-
-    fun getToken(): Flow<String> {
-        return prefs.getToken()
-    }
-
-}
